KAFKA-14683 Migrate #testOriginalTopicWithTopicMutatingTransformations to Mockito
